동작 수정 및 벤치마크 테스트 추가.

+++ BinaryXml/RawString.cs
@@ -96,7 +96,7 @@ namespace BinaryXml
             return TryToBoolean(out var value) ? value : throw new FormatException("Incorrect format.");
         }
 
-        public long ToInt()
+        public int ToInt()
         {
             return TryToInt(out var value) ? value : throw new FormatException("Incorrect format.");
         }
